Update getSuggestionsCard to use adaptive card and update qna dialog tests to use the new parameter
